MDS-5707 Cleanup
