+++ services/core-api/app/api/utils/setup_marshmallow.py
@@ -134,7 +134,7 @@ def setup_schema(Base, session):
                 except Exception as e:
                     raise e
 
-        for class_ in Base.registry._class_registry.values():
+        for class_ in list(Base.registry._class_registry.values()):
             if hasattr(class_, "_schema"):
                 try:
                     mapper = inspect(class_)
