+++ services/core-api/app/api/utils/static_data.py
@@ -15,7 +15,7 @@ Return: None
 
 
 def setup_static_data(Base):
-    for class_ in Base.registry._class_registry.values():
+    for class_ in list(Base.registry._class_registry.values()):
         if hasattr(class_, "__tablename__") or getattr(class_, "__create_schema__", False):
 
             try:
