+++ services/core-api/app/commands.py
@@ -17,6 +17,7 @@ from app.api.verifiable_credentials.models.credentials import PartyVerifiableCre
 from app.api.mines.permits.permit.models.permit import Permit
 
 from tests.factories import MineFactory, MinePartyAppointmentFactory, MinespaceSubscriptionFactory, MinespaceUserFactory, NOWSubmissionFactory, NOWApplicationIdentityFactory
+from .cli_commands.generate_history_table_migration import generate_history_table_migration
 
 
 def register_commands(app):
@@ -156,67 +157,12 @@ def register_commands(app):
 
     @app.cli.command('generate_history_table_migration')
     @click.argument('table')
-    def generate_history_table_migration(table):
+    def do_generate_history_table_command(table):
         """
         Generate a migration file that contains the history table definition for the specified table.
         Uses SQLAlchemy-continuum to generate the history table definition.
+
+        Example usage:
+            flask generate_history_table_migration mine_tailings_storage_facility
         """
-        
-        print("Generating history table definition for table:")
-        version_table_name = f'{table}_version'
-    
-        # Find the SQLAlchemy model for the table
-        table_class = next((c for c in db.Model.registry.mappers if c.class_.__tablename__ == table), None)
-
-        if table_class is None:
-            print(f'Could not find SQLAlchemy model for table {table}')
-            return
-
-        # Find the SQLAlchemy-continuum version table for the table
-        version_table_class = next((c for c in db.Model.registry.mappers if c.class_.__tablename__ == version_table_name), None)
-
-        if version_table_class is None:
-            print(f'Could not find SQLalchemy-continuum table definition for table {table}.')
-            print(f'Ensure that the {str(table_class.class_)} model has the __versioned__ = {{}} property defined.')
-            return
-
-        version_table_definition = version_table_class.class_.__table__
-        print(dir(version_table_definition))
-        dt = datetime.datetime.today().strftime('%Y.%m.%d.%H.%M')
-
-        migration_name = f'/migrations/V{dt}__add_{version_table_name}_history_table.sql'
-
-        with open(migration_name, 'w') as f:
-            f.write("-- This file was generated by the generate_history_table_ddl command\n")
-            f.write("-- The file contains the corresponding history table definition for the {table} table\n")
-            f.write(str(CreateTable(version_table_definition).compile(dialect=postgresql.dialect())).strip())
-            f.write(';\n')
-    
-           
-            for index in version_table_definition.indexes:
-                create_stmt2 = CreateIndex(index).compile(dialect=postgresql.dialect())
-                f.write(str(create_stmt2).strip())
-                f.write(';\n')
-
-        data_migdt = (datetime.datetime.today() + datetime.timedelta(minutes=1)).strftime('%Y.%m.%d.%H.%M')
-        migration_name = f'/migrations/V{data_migdt}__add_{version_table_name}_history_table_backfill.sql'
-
-        with open(migration_name, 'w') as f:
-            table_columns = []
-
-            for column in table_class.class_.__table__.columns:
-                table_columns.append('"%s"' % column.name)
-
-            insert_stmt = 'with transaction AS (insert into transaction(id) values(DEFAULT) RETURNING id)\n' \
-                    'insert into {version_table_name} (transaction_id, operation_type, end_transaction_id, {table_columns})\n'\
-                    'select t.id, \'0\', null, {table_columns} from {table},transaction t\n' \
-                    .format(
-                        table_columns=', '.join(table_columns),
-                        table=table,
-                        version_table_name=version_table_name
-                    )
-            
-            f.write("-- This file was generated by the generate_history_table_ddl command\n")
-            f.write("-- The file contains the data migration to backfill history records for the {table} table\n")
-            f.write(insert_stmt)
-            f.write(';\n')
+        generate_history_table_migration(table)
